+++ example/flutter_example/android/app/src/main/kotlin/dev/dhmgroup/flutter_example/MainActivity.kt
@@ -1,0 +1,6 @@
+package dev.dhmgroup.flutter_example
+
+import io.flutter.embedding.android.FlutterActivity
+
+class MainActivity: FlutterActivity() {
+}